requirements.txt
streamlit
google-generativeai
yt-dlp
moviepy
gtts
pillow
requests
numpy